fix(app): keep the settings sidebar mounted across sections
Every settings section is its own route, and the desktop sidebar was
rendered by the settings screen that navigation replaces. Selecting a
section therefore tore down the whole screen and mounted a fresh one, so
the sidebar came back scrolled to the top — losing the user's place
whenever they picked a section from below the fold.

Move the sidebar into a new SettingsChrome that wraps the root stack, so
navigation only swaps the detail pane. The route topology is unchanged:
docs/expo-router.md requires the settings routes to stay distinct
siblings, so the chrome resolves the selected section from the pathname
via parseSettingsViewFromPathname() instead of route params.

The add-host modals move with it, behind a small store, because the
sidebar trigger and the modals no longer share a component. The host
picker choice now survives section changes too.

Compact layout is untouched: it pushes a full-screen page per section,
where scrolling to the top is ordinary navigation.

Also converts the extracted sidebar off useUnistyles() to the
withUnistyles/uniProps pattern, since the ban's grandfathering does not
follow code into a new file.
