+++ packages/app/src/screens/settings/settings-sidebar.tsx
@@ -1,0 +1,510 @@
+import { useCallback, useMemo, useRef, useState } from "react";
+import type { ComponentType } from "react";
+import { Pressable, ScrollView, Text, View, type PressableStateCallbackType } from "react-native";
+import { useSafeAreaInsets } from "react-native-safe-area-context";
+import { StyleSheet, withUnistyles } from "react-native-unistyles";
+import { useTranslation } from "react-i18next";
+import {
+  ArrowLeft,
+  Settings,
+  Palette,
+  Server,
+  Network,
+  Bot,
+  Boxes,
+  Gauge,
+  Keyboard,
+  Stethoscope,
+  Info,
+  Bell,
+  Shield,
+  Puzzle,
+  Plus,
+  FolderGit2,
+  SquareTerminal,
+  Code2,
+  Smartphone,
+  Sparkles,
+  Blocks,
+  PanelsTopLeft,
+} from "lucide-react-native";
+import { ComboboxTrigger } from "@/components/ui/combobox-trigger";
+import { SidebarHeaderRow } from "@/components/sidebar/sidebar-header-row";
+import { SidebarSeparator } from "@/components/sidebar/sidebar-separator";
+import { HostPicker as SharedHostPicker } from "@/components/hosts/host-picker";
+import { HostStatusDot } from "@/components/host-status-dot";
+import { useHosts } from "@/runtime/host-runtime";
+import { orderHostsLocalFirst, type HostProfile } from "@/types/host-connection";
+import { TitlebarDragRegion } from "@/components/desktop/titlebar-drag-region";
+import { WindowChromeSafeArea } from "@/utils/desktop-window";
+import { isElectronRuntime } from "@/desktop/host";
+import { SETTINGS_DESKTOP_SIDEBAR_WIDTH } from "@/constants/layout";
+import { useLocalDaemonServerId } from "@/hooks/use-is-local-daemon";
+import {
+  type EnableBuiltInDaemonOption,
+  useEnableBuiltInDaemonOption,
+} from "@/desktop/hooks/use-enable-built-in-daemon-option";
+import type { HostSectionSlug, SettingsSectionSlug } from "@/utils/host-routes";
+import type { SettingsView } from "@/navigation/settings-navigation";
+import { isWeb } from "@/constants/platform";
+import { ICON_SIZE, type Theme } from "@/styles/theme";
+
+const foregroundColorMapping = (theme: Theme) => ({ color: theme.colors.foreground });
+const foregroundMutedColorMapping = (theme: Theme) => ({ color: theme.colors.foregroundMuted });
+const ThemedPlus = withUnistyles(Plus);
+const ThemedServer = withUnistyles(Server);
+
+type SidebarIcon = ComponentType<{ size: number; color: string }>;
+
+function createThemedSidebarIcon(Icon: SidebarIcon) {
+  return withUnistyles(Icon);
+}
+
+const themedSidebarIcons = new Map<SidebarIcon, ReturnType<typeof createThemedSidebarIcon>>();
+
+function getThemedSidebarIcon(Icon: SidebarIcon) {
+  const cached = themedSidebarIcons.get(Icon);
+  if (cached) return cached;
+  const themed = createThemedSidebarIcon(Icon);
+  themedSidebarIcons.set(Icon, themed);
+  return themed;
+}
+
+interface SidebarSectionItem {
+  id: SettingsSectionSlug;
+  labelKey: string;
+  icon: SidebarIcon;
+  desktopOnly?: boolean;
+  webOnly?: boolean;
+}
+
+export const SIDEBAR_SECTION_ITEMS: SidebarSectionItem[] = [
+  { id: "general", labelKey: "settings.sections.general", icon: Settings },
+  { id: "appearance", labelKey: "settings.sections.appearance", icon: Palette },
+  {
+    id: "layout",
+    labelKey: "settings.sections.layout",
+    icon: PanelsTopLeft,
+    desktopOnly: true,
+  },
+  { id: "editor", labelKey: "settings.sections.editor", icon: Code2, webOnly: true },
+  { id: "shortcuts", labelKey: "settings.sections.shortcuts", icon: Keyboard, desktopOnly: true },
+  {
+    id: "integrations",
+    labelKey: "settings.sections.integrations",
+    icon: Puzzle,
+    desktopOnly: true,
+  },
+  {
+    id: "notifications",
+    labelKey: "settings.sections.notifications",
+    icon: Bell,
+    desktopOnly: true,
+  },
+  {
+    id: "permissions",
+    labelKey: "settings.sections.permissions",
+    icon: Shield,
+    desktopOnly: true,
+  },
+  { id: "diagnostics", labelKey: "settings.sections.diagnostics", icon: Stethoscope },
+  { id: "about", labelKey: "settings.sections.about", icon: Info },
+];
+
+interface HostSectionItem {
+  id: HostSectionSlug;
+  labelKey: string;
+  icon: SidebarIcon;
+}
+
+export const HOST_SECTION_ITEMS: HostSectionItem[] = [
+  { id: "host", labelKey: "settings.hostSections.host", icon: Server },
+  { id: "projects", labelKey: "settings.hostSections.projects", icon: FolderGit2 },
+  { id: "connections", labelKey: "settings.hostSections.connections", icon: Network },
+  { id: "pair-device", labelKey: "openProject.tiles.pairDevice.title", icon: Smartphone },
+  { id: "agents", labelKey: "settings.hostSections.agents", icon: Bot },
+  { id: "metadata", labelKey: "settings.hostSections.metadata", icon: Sparkles },
+  { id: "workspaces", labelKey: "settings.hostSections.workspaces", icon: FolderGit2 },
+  { id: "providers", labelKey: "settings.hostSections.providers", icon: Boxes },
+  { id: "usage", labelKey: "settings.hostSections.usage", icon: Gauge },
+  { id: "terminals", labelKey: "settings.hostSections.terminals", icon: SquareTerminal },
+  { id: "plugins", labelKey: "settings.hostSections.plugins", icon: Blocks },
+];
+
+function sidebarItemStyle({ hovered }: PressableStateCallbackType & { hovered?: boolean }) {
+  return [sidebarStyles.item, Boolean(hovered) && sidebarStyles.itemHovered];
+}
+
+function selectedSidebarItemStyle({ hovered }: PressableStateCallbackType & { hovered?: boolean }) {
+  return [
+    sidebarStyles.item,
+    Boolean(hovered) && sidebarStyles.itemHovered,
+    sidebarStyles.itemSelected,
+  ];
+}
+
+/**
+ * Local daemon first, then remaining hosts in their existing order.
+ */
+export function useSortedHosts(hosts: HostProfile[], localServerId: string | null): HostProfile[] {
+  return useMemo(() => orderHostsLocalFirst(hosts, localServerId), [hosts, localServerId]);
+}
+
+interface SidebarNavButtonProps<Id extends string> {
+  itemId: Id;
+  label: string;
+  icon: SidebarIcon;
+  isSelected: boolean;
+  onSelect: (id: Id) => void;
+  testID?: string;
+}
+
+function SidebarNavButton<Id extends string>({
+  itemId,
+  label,
+  icon: IconComponent,
+  isSelected,
+  onSelect,
+  testID,
+}: SidebarNavButtonProps<Id>) {
+  const ThemedIcon = getThemedSidebarIcon(IconComponent);
+  const handlePress = useCallback(() => {
+    onSelect(itemId);
+  }, [onSelect, itemId]);
+  const accessibilityState = useMemo(() => ({ selected: isSelected }), [isSelected]);
+  return (
+    <Pressable
+      accessibilityRole="button"
+      accessibilityState={accessibilityState}
+      onPress={handlePress}
+      testID={testID}
+      style={isSelected ? selectedSidebarItemStyle : sidebarItemStyle}
+    >
+      <ThemedIcon
+        size={ICON_SIZE.md}
+        uniProps={isSelected ? foregroundColorMapping : foregroundMutedColorMapping}
+      />
+      <Text
+        style={[sidebarStyles.label, isSelected && sidebarStyles.labelSelected]}
+        numberOfLines={1}
+      >
+        {label}
+      </Text>
+    </Pressable>
+  );
+}
+
+interface HostPickerProps {
+  activeServerId: string | null;
+  sortedHosts: HostProfile[];
+  onSelectHost: (serverId: string) => void;
+  onAddHost: () => void;
+  enableBuiltInDaemonOption: EnableBuiltInDaemonOption;
+}
+
+/**
+ * Scopes the host sections to a host. Reuses the canonical sidebar host
+ * switcher pattern (left-sidebar.tsx): a quiet row-styled trigger opening a
+ * <Combobox>. The local host is listed first, each row shows the connection it
+ * is using right now; an "Add host" row is always reachable from the list —
+ * even with a single host.
+ */
+function HostPicker({
+  activeServerId,
+  sortedHosts,
+  onSelectHost,
+  onAddHost,
+  enableBuiltInDaemonOption,
+}: HostPickerProps) {
+  const { t } = useTranslation();
+  const [isOpen, setIsOpen] = useState(false);
+  const triggerRef = useRef<View | null>(null);
+  const activeHost =
+    sortedHosts.find((host) => host.serverId === activeServerId) ?? sortedHosts[0] ?? null;
+
+  const handleOpen = useCallback(() => setIsOpen(true), []);
+  const hostOptionTestID = useCallback(
+    (serverId: string) => `settings-host-picker-item-${serverId}`,
+    [],
+  );
+  const triggerStyle = useCallback(
+    ({ hovered = false }: PressableStateCallbackType & { hovered?: boolean }) => [
+      sidebarStyles.pickerTrigger,
+      hovered && sidebarStyles.pickerTriggerHovered,
+    ],
+    [],
+  );
+
+  return (
+    <SharedHostPicker
+      hosts={sortedHosts}
+      value={activeServerId ?? ""}
+      onSelect={onSelectHost}
+      open={isOpen}
+      onOpenChange={setIsOpen}
+      anchorRef={triggerRef}
+      includeAddHost
+      onAddHost={onAddHost}
+      includeEnableBuiltInDaemon={enableBuiltInDaemonOption.visible}
+      onEnableBuiltInDaemon={enableBuiltInDaemonOption.onPress}
+      showActiveConnection
+      searchable={false}
+      title={t("settings.hostPicker.switchHost")}
+      desktopMinWidth={240}
+      addHostTestID="settings-add-host"
+      hostOptionTestID={hostOptionTestID}
+    >
+      <ComboboxTrigger
+        ref={triggerRef}
+        block
+        style={triggerStyle}
+        onPress={handleOpen}
+        accessibilityRole="button"
+        accessibilityLabel={t("settings.hostPicker.switchHost")}
+        testID="settings-host-picker"
+      >
+        {activeHost ? (
+          <View style={sidebarStyles.pickerTriggerDot}>
+            <HostStatusDot serverId={activeHost.serverId} />
+          </View>
+        ) : null}
+        <Text style={sidebarStyles.pickerTriggerLabel} numberOfLines={1}>
+          {activeHost?.label ?? t("settings.groups.host")}
+        </Text>
+      </ComboboxTrigger>
+    </SharedHostPicker>
+  );
+}
+
+interface SettingsSidebarProps {
+  view: SettingsView;
+  onSelectSection: (section: SettingsSectionSlug) => void;
+  onSelectHostSection: (section: HostSectionSlug) => void;
+  onSelectHost: (serverId: string) => void;
+  onAddHost: () => void;
+  onBackToWorkspace: () => void;
+  activeHostServerId: string | null;
+  layout: "desktop" | "mobile";
+}
+
+export function SettingsSidebar({
+  view,
+  onSelectSection,
+  onSelectHostSection,
+  onSelectHost,
+  onAddHost,
+  onBackToWorkspace,
+  activeHostServerId,
+  layout,
+}: SettingsSidebarProps) {
+  const { t } = useTranslation();
+  const hosts = useHosts();
+  const localServerId = useLocalDaemonServerId();
+  const sortedHosts = useSortedHosts(hosts, localServerId);
+  const hasHosts = sortedHosts.length > 0;
+  const enableBuiltInDaemonOption = useEnableBuiltInDaemonOption();
+  const isDesktopApp = isElectronRuntime();
+  const items = SIDEBAR_SECTION_ITEMS.filter(
+    (item) => (!item.desktopOnly || isDesktopApp) && (!item.webOnly || isWeb),
+  );
+  const insets = useSafeAreaInsets();
+  const isDesktop = layout === "desktop";
+  const outerContainerStyle = useMemo(
+    () => [isDesktop ? sidebarStyles.desktopContainer : sidebarStyles.mobileContainer],
+    [isDesktop],
+  );
+  const innerContainerStyle = useMemo(
+    () => [{ flex: 1 }, isDesktop ? { paddingTop: insets.top } : null],
+    [insets.top, isDesktop],
+  );
+  const selectedSectionId = view.kind === "section" ? view.section : null;
+  let selectedHostSection: HostSectionSlug | null = null;
+  if (view.kind === "host") selectedHostSection = view.section;
+  if (view.kind === "project") selectedHostSection = "projects";
+  if (view.kind === "plugin") selectedHostSection = "plugins";
+
+  const sidebarBody = (
+    <>
+      <View style={sidebarStyles.list}>
+        <Text style={sidebarStyles.groupLabel}>{t("settings.groups.app")}</Text>
+        {items.map((item) => (
+          <SidebarNavButton
+            key={item.id}
+            itemId={item.id}
+            label={t(item.labelKey)}
+            icon={item.icon}
+            isSelected={selectedSectionId === item.id}
+            onSelect={onSelectSection}
+          />
+        ))}
+      </View>
+      <SidebarSeparator />
+      {hasHosts ? (
+        <View style={sidebarStyles.list}>
+          <Text style={sidebarStyles.groupLabel}>{t("settings.groups.host")}</Text>
+          <HostPicker
+            activeServerId={activeHostServerId}
+            sortedHosts={sortedHosts}
+            onSelectHost={onSelectHost}
+            onAddHost={onAddHost}
+            enableBuiltInDaemonOption={enableBuiltInDaemonOption}
+          />
+          {HOST_SECTION_ITEMS.map((item) => (
+            <SidebarNavButton
+              key={item.id}
+              itemId={item.id}
+              label={t(item.labelKey)}
+              icon={item.icon}
+              isSelected={selectedHostSection === item.id}
+              onSelect={onSelectHostSection}
+              testID={`settings-host-section-${item.id}`}
+            />
+          ))}
+        </View>
+      ) : (
+        <View style={sidebarStyles.list}>
+          <Pressable
+            accessibilityRole="button"
+            accessibilityLabel={t("settings.addHost")}
+            onPress={onAddHost}
+            testID="settings-add-host"
+            style={sidebarItemStyle}
+          >
+            <ThemedPlus size={ICON_SIZE.md} uniProps={foregroundMutedColorMapping} />
+            <Text style={sidebarStyles.label} numberOfLines={1}>
+              {t("settings.addHost")}
+            </Text>
+          </Pressable>
+          {enableBuiltInDaemonOption.visible ? (
+            <Pressable
+              accessibilityRole="button"
+              accessibilityLabel={t("settings.enableBuiltInDaemon")}
+              onPress={enableBuiltInDaemonOption.onPress}
+              testID="settings-enable-built-in-daemon"
+              style={sidebarItemStyle}
+            >
+              <ThemedServer size={ICON_SIZE.md} uniProps={foregroundMutedColorMapping} />
+              <Text style={sidebarStyles.label} numberOfLines={1}>
+                {t("settings.enableBuiltInDaemon")}
+              </Text>
+            </Pressable>
+          ) : null}
+        </View>
+      )}
+    </>
+  );
+
+  return (
+    <View
+      accessibilityLabel={t("settings.title")}
+      role="navigation"
+      style={outerContainerStyle}
+      testID="settings-sidebar"
+    >
+      {isDesktop ? (
+        <View style={innerContainerStyle}>
+          <View style={sidebarStyles.sidebarDragArea}>
+            <TitlebarDragRegion />
+            <WindowChromeSafeArea placement="below" />
+            <SidebarHeaderRow
+              icon={ArrowLeft}
+              label={t("settings.backToWorkspace")}
+              onPress={onBackToWorkspace}
+              testID="settings-back-to-workspace"
+            />
+          </View>
+          <ScrollView
+            style={sidebarStyles.scrollBody}
+            showsVerticalScrollIndicator={false}
+            testID="settings-sidebar-scroll-body"
+          >
+            {sidebarBody}
+          </ScrollView>
+        </View>
+      ) : (
+        sidebarBody
+      )}
+    </View>
+  );
+}
+
+const sidebarStyles = StyleSheet.create((theme) => ({
+  desktopContainer: {
+    width: SETTINGS_DESKTOP_SIDEBAR_WIDTH,
+    borderRightWidth: 1,
+    borderRightColor: theme.colors.border,
+    backgroundColor: theme.colors.surfaceSidebar,
+  },
+  scrollBody: {
+    flex: 1,
+  },
+  sidebarDragArea: {
+    position: "relative",
+  },
+  mobileContainer: {
+    paddingVertical: theme.spacing[2],
+    paddingHorizontal: theme.spacing[2],
+  },
+  list: {
+    paddingVertical: theme.spacing[2],
+    paddingHorizontal: theme.spacing[2],
+    gap: theme.spacing[1],
+  },
+  groupLabel: {
+    fontSize: theme.fontSize.base,
+    fontWeight: theme.fontWeight.medium,
+    color: theme.colors.foregroundMuted,
+    paddingHorizontal: theme.spacing[2],
+    paddingVertical: theme.spacing[1],
+  },
+  item: {
+    flexDirection: "row",
+    alignItems: "center",
+    gap: theme.spacing[2],
+    minHeight: 36,
+    paddingVertical: theme.spacing[2],
+    paddingHorizontal: theme.spacing[2],
+    borderRadius: theme.borderRadius.lg,
+  },
+  itemHovered: {
+    backgroundColor: theme.colors.surfaceSidebarHover,
+  },
+  itemSelected: {
+    backgroundColor: theme.colors.surfaceSidebarHover,
+  },
+  label: {
+    fontSize: theme.fontSize.base,
+    color: theme.colors.foregroundMuted,
+    fontWeight: theme.fontWeight.normal,
+    flex: 1,
+  },
+  labelSelected: {
+    color: theme.colors.foreground,
+  },
+  pickerTrigger: {
+    flexDirection: "row",
+    alignItems: "center",
+    gap: theme.spacing[2],
+    minHeight: 36,
+    paddingVertical: theme.spacing[2],
+    paddingHorizontal: theme.spacing[2],
+    borderRadius: theme.borderRadius.lg,
+  },
+  pickerTriggerHovered: {
+    backgroundColor: theme.colors.surfaceSidebarHover,
+  },
+  pickerTriggerLabel: {
+    flex: 1,
+    minWidth: 0,
+    fontSize: theme.fontSize.base,
+    color: theme.colors.foreground,
+    fontWeight: theme.fontWeight.normal,
+  },
+  // Match the setting items' icon footprint so the host label aligns with them.
+  pickerTriggerDot: {
+    width: theme.iconSize.md,
+    height: theme.iconSize.md,
+    alignItems: "center",
+    justifyContent: "center",
+  },
+}));
